third_party/blink/web_tests/external/wpt/webrtc/simulcast/negotiation-encodings.https-expected.txt

This is a testharness.js-based test.
[FAIL] addTrack, then sRD(simulcast recv offer) results in simulcast
  assert_equals: expected 1 but got 2
[FAIL] simulcast is not supported for audio
  assert_equals: expected 1 but got 2
[FAIL] sRD(recv simulcast answer) can narrow the simulcast envelope specified by addTransceiver
  assert_array_equals: expected property 0 to be 2 but got undefined (expected array [2] got [undefined])
[FAIL] sRD(recv simulcast answer) can narrow the simulcast envelope from a previous negotiation
  assert_array_equals: expected property 0 to be 2 but got undefined (expected array [2] got [undefined])
[FAIL] sRD(simulcast offer) can narrow the simulcast envelope from a previous negotiation
  assert_array_equals: [[SendEncodings]] is not updated in have-remote-offer for reoffers lengths differ, expected array ["foo", "bar"] length 2, got ["foo"] length 1
[FAIL] Duplicate rids in sRD(offer) are ignored
  assert_equals: expected 1 but got 2
[FAIL] Choices in rids in sRD(offer) are ignored
  assert_equals: expected 1 but got 2
[FAIL] addTrack, then rollback of sRD(simulcast offer), brings us back to having a single encoding without a rid
  assert_equals: expected 1 but got 2
[FAIL] sRD(simulcast offer), addTrack, then rollback brings us back to having a single encoding
  assert_array_equals: lengths differ, expected array [undefined] length 1, got ["foo", "bar"] length 2
[FAIL] Rollback of sRD(reoffer) with a single rid results in all previous encodings
  assert_array_equals: lengths differ, expected array ["foo", "bar"] length 2, got ["bar"] length 1
[FAIL] sRD(recv simulcast answer) can narrow the simulcast envelope specified by addTransceiver by removing the first encoding
  assert_array_equals: expected property 0 to be 1 but got undefined (expected array [1] got [undefined])
[FAIL] sRD(recv simulcast answer) can narrow the simulcast envelope from a previous negotiation by removing the first encoding
  assert_array_equals: expected property 0 to be 1 but got undefined (expected array [1] got [undefined])
[FAIL] sRD(simulcast offer) can narrow the simulcast envelope from a previous negotiation by removing the first encoding
  assert_array_equals: [[SendEncodings]] is not updated in have-remote-offer for reoffers lengths differ, expected array ["foo", "bar"] length 2, got ["bar"] length 1
Harness: the test ran to completion.
